PREFAB - Player Character Complete (again)
Worked with Lucas, Finlay, and Niko to find a good camera angle that fits their vision. I also fixed the movement controller script, making it so that the player will properly move in the direction of the pressed key. Lastly, made it so that the camera is capable of following the player as they move.