HUDControllerClass Complete
HUD Controller Class has been completed for the dialogue HUD. This lets the player enter into dialogue, and select from 4 different options. Based on the option selected, the next dialogue is loaded. For the time being, the dialogue HUD can be accessed with the "K" key, and closed with the "L" key.