InventoryHUD Updated
We created a folder and changed TestStick to StandardInventoryItem to allow for inventory items to be created much easier. We created a rock as the testing object. MovementController and HUDController were slightly changed to accommodate this change. The inventoryhud should be working once the buttons in the inventory canvas are slightly changed to have images instead of rawimages.