InventoryHUDController
An empty StandardInventoryItem was created for empty inventory slots. The new canvas with images instead of rawimages was added. The MovementController can now add empty items through TestKeys, and properly prints the amount of items in the inventory. Lastly, the InventoryManager script was optimized for different inventoryTypes, and no longer allows for items to be added above the limit.