InventoryHUDController Complete
Created prefabs to make using the GameManager easier. The proper prefab is UI_GameManagerAssests in the UI Prefab folder. The player and player inventory must still be added to properly access the player's inventory. Y adds sticks to the player's inventory, 2 closes the inventory, and clicking on an inventory item will show more information about it.