HUDFollowerUpdated
The updateFollowers() method should work properly and should allow the player to access either the Player's follower list or the HomeBase follower list. FollowerManager was completely overhauled to now work properly. PlayerController has some methods that need to be fixed regarding adding and removing NPC's stats when following the player.